hard refactoring in progress
