+++ app/src/main/AndroidManifest.xml
@@ -21,9 +21,9 @@
             android:name="com.alexm.chess960.RulesActivity"
             android:parentActivityName="com.alexm.chess960.MainMenu" />
         <activity
-            android:name="com.alexm.chess960.randomPack.RandomActivity"
+            android:name="com.alexm.chess960.randompos.RandomActivity"
             android:parentActivityName="com.alexm.chess960.MainMenu" />
-        <activity android:name="com.alexm.chess960.clockPack.ClockActivity"
+        <activity android:name="com.alexm.chess960.clockpack.ClockActivity"
             android:theme="@android:style/Theme.Black.NoTitleBar.Fullscreen"/>
     </application>
