+++ app/src/main/java/com/alexm/chess960/randompos/IRandomPos.kt
@@ -1,4 +1,4 @@
-package com.alexm.chess960.interfaces
+package com.alexm.chess960.randompos
 
 /**
  * Created by alexm on 21/05/2018.
